Extract AbstractGenerator class
